+++ src/num/range.rs
@@ -1,7 +1,7 @@
 #[derive(Clone, Copy, Debug)]
 pub struct Range {
-    pub lo: f32,
-    pub hi: f32,
+    pub lo: f64,
+    pub hi: f64,
 }
 
 impl Range {
@@ -12,7 +12,7 @@ impl Range {
         }
     }
 
-    pub fn bounds(self) -> (f32, f32) {
+    pub fn bounds(self) -> (f64, f64) {
         (self.lo, self.hi)
     }
 
@@ -31,8 +31,8 @@ macro_rules! impl_from {
             impl From<std::ops::Range<$ty>> for Range {
                 fn from(r: std::ops::Range<$ty>) -> Self {
                     Self {
-                        lo: r.start as f32,
-                        hi: r.end as f32,
+                        lo: r.start as f64,
+                        hi: r.end as f64,
                     }
                 }
             }
@@ -40,4 +40,4 @@ macro_rules! impl_from {
     };
 }
 
-impl_from!(f32, i32);
+impl_from!(f64, i32);
